LayoutTests/http/tests/xmlhttprequest/get-dangerous-headers-expected.txt

CONSOLE MESSAGE: line 1: Refused to get unsafe header "Set-Cookie"
CONSOLE MESSAGE: line 1: Refused to get unsafe header "set-cookie"
CONSOLE MESSAGE: line 1: Refused to get unsafe header "Set-Cookie2"
CONSOLE MESSAGE: line 1: Refused to get unsafe header "set-cookie2"
Test that getResponseHeader and getAllResponseHeaders cannot be used to get the cookie header fields.

PASS
